requirements.txt
fastapi[standard]
pip-tools
snowflake-connector-python
snowflake-sqlalchemy
sqlmodel